Add feat-units-sections, refactoring and adjust responsiveness of small screens
