+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Navbar from "./_components/Navbar";
+import Navbar from "./_components/shared/Navbar";
 
 const inter = Inter({ subsets: ["latin"] });
